Convert to std::array and std::string. (libmyth)

+++ mythtv/libs/libmyth/audio/audiooutput.h
@@ -152,7 +152,7 @@ class MPUBLIC AudioOutput : public VolumeBase, public OutputListeners
     //  Only really used by the AudioOutputNULL object
     virtual void bufferOutputData(bool y) = 0;
     virtual int readOutputData(unsigned char *read_buffer,
-                               int max_length) = 0;
+                               size_t max_length) = 0;
 
     virtual bool IsUpmixing(void)   { return false; }
     virtual bool ToggleUpmix(void)  { return false; }
